requirements.txt
Orange3
resdk
anyqt
pyqt5